# Conflicts:
#	lib/ui/screens/order_successful_screen.dart
#	lib/ui/screens/planogram_product_screen.dart
#	lib/ui/screens/shipment_verification_screen.dart
#	lib/ui/screens/store_category_screen.dart
#	lib/ui/widget/common_product_details_widget.dart
#	pubspec.lock

+++ lib/bloc/order_successful/order_successful_state.dart
@@ -4,11 +4,11 @@ part of 'order_successful_bloc.dart';
 @freezed
 class OrderSuccessfulState with _$OrderSuccessfulState {
   const factory OrderSuccessfulState({
-    required int totalCredit,
-    required int thisMonthExpense,
-    required int lastMonthExpense,
+    required double totalCredit,
+    required double thisMonthExpense,
+    required double lastMonthExpense,
     required int orderThisMonth,
-    required int balance,
+    required double balance,
   }) = _OrderSuccessfulState;
 
   factory OrderSuccessfulState.initial() =>
