+++ lib/data/model/order_model/product_details_model.dart
@@ -4,7 +4,7 @@ class ProductDetailsModel{
    String? productName ;
    String? mainImage;
    int? totalQuantity;
-   int? totalPayment;
+   double? totalPayment;
    bool isProductIssue;
    bool isDelete;
   String cartProductId;
